requirements.txt
pandas>=2.0.0
numpy>=1.24.0
psutil>=5.9.0
jupyter>=1.0.0
notebook>=7.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
sodapy
python-dotenv
openmeteo_requests
requests_cache
retry_requests
pyspark